Update repository URL in README
The commit updates the repository URL in the README file. The previous
URL was incorrectly mentioned as
"https://github.com/yourusername/airogit.git". The correct repository
URL is now mentioned as "https://github.com/CocoFruit/AIROGIT.git".

This change was made to ensure that users can access the correct
repository location when cloning or downloading the project.